refactor: split init into init and join, require --key on both
Split the overloaded init command into separate init (bootstrap) and
join subcommands. Both require --name, --repo, and --key. The init
command exports the git-crypt key to the user-specified --key path
instead of a hardcoded parent directory. The join command uses --key
to unlock the repo.

Also: validate host names, fix env.template docs example, use
@{upstream} in diff instead of hardcoded origin/main, remove dead
code.
