+++ cmd/gitops/gitops.go
@@ -34,6 +34,7 @@ git-crypt, and multi-server deployments with push/pull workflows.`,
 	}
 
 	gitopsCmd.AddCommand(newInitCmd(&instance))
+	gitopsCmd.AddCommand(newJoinCmd(&instance))
 	gitopsCmd.AddCommand(newPushCmd(&instance))
 	gitopsCmd.AddCommand(newPullCmd(&instance))
 	gitopsCmd.AddCommand(newStatusCmd(&instance))
